tools/requirements.txt
fastapi
uvicorn[standard]
transformers
Pillow
torch
sentencepiece
tokenizers
